// lib/types.ts
export type DaggerHeartAbility = {
  type: 'daggerheart';
  label: string;
  modifier: number;
  list: string[];
};

export type DNDAbility = {
  type: 'dnd';
  label: string;
  total: number;
  modifier: number;
  isProficient: boolean;
  savingThrow: number;
}